refactor(openrouter): extract dedicated crate from openai
OpenRouter-specific code was accumulating inside the everruns-openai
crate (and the vendor-neutral Open Responses driver in core). Move the
OpenRouter driver, model discovery, and model-profile types into a new
everruns-openrouter crate, mirroring the per-provider crate layout used
by anthropic/gemini/bedrock.

- New crate wraps core's OpenResponsesProtocolChatDriver tagged with
  DriverId::OpenRouter, same as before; behavior is unchanged.
- Generic /models discovery helpers (url_host_eq, normalize_api_url,
  models_url_for_api_url, the bearer/Azure auth helper, status error)
  are promoted into core::openai_protocol so both provider crates reuse
  them instead of duplicating.
- OpenRouter registration moves to everruns_openrouter::register_driver;
  registration call sites (worker, llm-tests) now call it.

OpenRouter request-body augmentation in core (routing/plugins/session_id)
is intentionally left for a follow-up phase.
